# openapi.yaml
openapi: 3.1.0
info:
  title: SILAGO API
  description: |
    Marketplace kampus Polines. Base prefix `/api/v1`. Auth via Laravel Sanctum Bearer token.
    Pagination default 15 (20 untuk conversations/notifications). Locale `id` (Carbon `id`).
    Storage URL via `Storage::url()` (`/storage/...`).
  version: 1.0.0
  contact:
    name: SILAGO
servers:
  - url: http://localhost:8000/api/v1
    description: Local
  - url: https://silago.example.com/api/v1
    description: Production (ganti host)
security:
  - bearerAuth: []
...
paths:
...
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email dengan kode 6 digit
      security: []
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, example: "123456" }
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Email berhasil diverifikasi" }
                  user: { $ref: "#/components/schemas/UserResource" }
        '400':
          description: Kode tidak valid / kedaluwarsa (cache 15 menit)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Kirim ulang kode verifikasi (rate limit 1/60s)
      security: []
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { message: { type: string, example: "Kode verifikasi telah dikirim ulang." } } } } } }
        '400': { description: Sudah terverifikasi }
        '429': { $ref: "#/components/responses/TooManyRequests" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Sanctum personal_access_token. Header: `Authorization: Bearer <token>`"
...
  responses:
...
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { message: "Terlalu banyak permintaan. Coba lagi nanti." }
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
    UserResource:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        nim: { type: string, nullable: true }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        profile_photo_url: { type: string, nullable: true, format: uri }
        role: { type: string, enum: [user, moderator, admin] }
        status: { type: string, enum: [active, suspended] }
        email_verified_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...